fix reviewed poinst

+++ src/include/script_runtime.hpp
@@ -204,10 +204,10 @@ namespace cshort
 
 
     struct ScriptRunner {
-        static int runScriptWithLength(const char* script, int byteLength);
+        static int64_t runScriptWithLength(const char* script, int byteLength);
 
         template<std::size_t SIZE>
-        static int runScript(const char(&text)[SIZE])
+        static int64_t runScript(const char(&text)[SIZE])
         {
             return runScriptWithLength(text, SIZE - 1);
         }
